requirements.txt
argparse
black
eradicate
fasttext
Flask
flask-cors
jsonlines
nltk
numpy
python-Levenshtein
requests
rouge
scikit-learn
sentencepiece
simhash
torch
tqdm
transformers
wandb